# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admin panel API",
        "description": "Codeyad-React Learning-Ecommerce admin panel API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/admin/brands": {
            "get": {
                "tags": [
                    "Brands"
                ],
                "summary": "Get all brands",
                "description": "Get all brands",
                "operationId": "allBrands",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "[{}, {}]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Brands"
                ],
                "summary": "Add brand",
                "description": "Store one brand",
                "operationId": "addBrand",
                "requestBody": {
                    "description": "Add one brand",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "original_name"
                                ],
                                "properties": {
                                    "original_name": {
                                        "type": "string",
                                        "example": "example name"
                                    },
                                    "persian_name": {
                                        "type": "string",
                                        "example": "نام تست"
                                    },
                                    "descriptions": {
                                        "type": "string",
                                        "example": "example description"
                                    },
                                    "logo": {
                                        "description": "file to upload",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "[{}, {}]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/admin/brands/{id}": {
            "get": {
                "tags": [
                    "Brands"
                ],
                "summary": "Get one brand",
                "description": "Get one brand with id",
                "operationId": "oneBrand",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "{}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Brands"
                ],
                "summary": "Update brand",
                "description": "Edit one brand",
                "operationId": "editBrand",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Edit one brand",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "original_name"
                                ],
                                "properties": {
                                    "original_name": {
                                        "type": "string",
                                        "example": "example name"
                                    },
                                    "persian_name": {
                                        "type": "string",
                                        "example": "نام تست"
                                    },
                                    "descriptions": {
                                        "type": "string",
                                        "example": "example description"
                                    },
                                    "logo": {
                                        "description": "file to upload",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "{...}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Brands"
                ],
                "summary": "Delete brand",
                "description": "Delete one brand",
                "operationId": "deleteBrend",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "حذف با موفقیت انجام شد"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}